+++ RythmGame.Game/Run/Run.cpp
@@ -2,7 +2,7 @@
 
 std::chrono::high_resolution_clock::time_point RythmGame::Game::Utils::offsetTime;
 
-long  RythmGame::Game::Utils::deltaTime;
+float  RythmGame::Game::Utils::deltaTime;
 float RythmGame::Game::Utils::velocity;
 bool  RythmGame::Game::Utils::isHorizontal;
